ippgtoolbox/detection/skin/deeplab/deeplab_MobileNetV2/requirements.txt
keras==2.2.4
tensorflow==2.5.1
Pillow
progressbar2
numpy
matplotlib
tqdm
opencv-python